mv talk and experiments up
